==> scripts/draft5/0_constants.R <==
## Set up constants used across multiple scripts
##
## 2022-11-10
## Peter Regier
##
# ########### #
# ########### #

# Constants --------------------------------------------------------------------

## Set ggplot theme
theme_set(theme_bw())

## Set seasonal color scheme
season_colors = c("#595F72", "#46ACC2")
